chore(encode,render): remove leftover native webp logic
